+++ projects/canviewer/BUILD
@@ -9,13 +9,12 @@ py_binary(
     data = [
         "//vehicle/mkv:mkv.dbc",
         ":config.yml",
-        ":logo.png",
-        ":style.qss",
     ],
     main = "canviewer.py",
     visibility = ["//visibility:public"],
     deps = [
         ":canserver",
+        "//projects/canviewer/gui",
     ],
 )
